Use HTML in FAQ answers
This changes the `jest.config.ts` to support importing `lit-html`.

Prior to this, jest would choke on test files that included `lit-html`
(transitively). This prevented us from using HTML in the FAQ questions,
since those questions are used in the test suite.

This includes some other changes:

* Since the questions can now be in a file that imports `lit-html`, they
  are moved to `faq/index.ts` to simplify the file tree,
* The Windows Hello now has a direct (HTML) link, as opposed to only
  using the `links` property of `Question`,
* The `Question` type has added doc,
* The behavior of creating a safe `<a>` tag is extracted in
  `mkExternalLink` and reused where applicable in the FAQ.
